// web/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        ProblemDetails: {
            /** Detail */
            detail: string;
            /**
             * Errors
             * @default null
             */
            errors: {
                [key: string]: string[];
            } | null;
            /** Status */
            status: number;
            /** Title */
            title: string;
            /** Traceid */
            traceId: string;
            /** Type */
            type: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}